refactor: run sveld before tsc for blade-svelte type generation
sveld emits per-component .svelte.d.ts files; tsc then writes the real
components barrel without touching those stubs. Drop the backup/restore
workaround and validate the published types after both steps complete.
