+++ .changeset/blade-svelte-types-barrel-fix.md
@@ -2,4 +2,4 @@
 "@razorpay/blade-svelte": patch
 ---
 
-fix: restore components type barrel after sveld overwrites tsc-generated index.d.ts
+fix: run tsc after sveld so the components type barrel is not overwritten
