+++ packages/blade-svelte/src/components/sveld-entry.js
@@ -1,5 +1,6 @@
 /**
  * sveld entry for glob-based `.svelte.d.ts` generation.
- * Glob mode discovers all `.svelte` files under this directory; this file exists only to satisfy sveld's entry requirement.
+ * Glob mode discovers all `.svelte` files under this directory. Exports are empty on purpose —
+ * tsc emits the real components barrel from index.ts after sveld runs.
  */
 export {};
